WIP - Move Miner and Consensus out
